+++ lib/manager/npm/extract/index.js
@@ -119,7 +119,7 @@ async function extractDependencies(content, packageFile, config) {
           if (depType === 'engines') {
             if (depName === 'node') {
               dep.purl = 'pkg:github/nodejs/node?clean=true';
-              dep.versionScheme = 'nodever';
+              dep.versionScheme = 'node';
             } else if (depName === 'yarn') {
               dep.purl = 'pkg:npm/yarn';
               dep.commitMessageTopic = 'Yarn';
