+++ lib/manager/nvm/extract.js
@@ -9,7 +9,7 @@ function extractDependencies(content) {
     depName: 'node',
     currentValue: content.trim(),
     purl: 'pkg:github/nodejs/node?clean=true',
-    versionScheme: 'nodever',
+    versionScheme: 'node',
   };
   if (!semver.isValid(dep.currentValue)) {
     dep.skipReason = 'unsupported-version';
