+++ lib/manager/travis/extract.js
@@ -13,7 +13,7 @@ function extractDependencies(content) {
       {
         depName: 'node',
         currentValue: doc.node_js,
-        versionScheme: 'nodever',
+        versionScheme: 'node',
       },
     ];
   }
