+++ lib/versioning/index.js
@@ -1,13 +1,13 @@
 const semver = require('./semver');
 const semverComposer = require('./semver-composer');
 const pep440 = require('./pep440');
-const nodever = require('./nodever');
+const node = require('./node');
 
 const schemes = {
   semver,
   semverComposer,
   pep440,
-  nodever,
+  node,
 };
 
 module.exports = function getVersionScheme(versionScheme) {
